[HelpMenuList] Minor optimisations
Clean up/simplify the calculation of the list of buttons that apply
to an ActionMap entry. This is the innermost loop of the construction
of the help display list.

Also tidy up _mergeButLists() and make it a little faster (using
sets to do the same thing is slower that the new code on the short
lists that actually occur in the code).