Fix up title decorations in channel selection list
The screen title in the channel selection screen is decorated with
various mode markers, e.g. "MyBouquet (TV) [move mode]".

The code that manipulates thes parkers is unreliable. For example,
the bouquet "## ( MUSIC & NEWS ) ##" is marked as "## ( MUSIC &
NEWS ) [move mode]  ## (TV)" in move mode, rather than the expected
"## ( MUSIC & NEWS )  ## (TV) [move mode]".

Also, if PiP is showing and "Zap focus to Picture in Picture" is
set, then the bouquet name can be doubled up in the title of the
channel list screen, e.g: "* 16,0E A1 HR* 16,0E A1 HR (PiP)".

This code rewrites buildTitleString() to build title string with
the appropriate decorations using the underlying state variabes in
the classes. It is just called whenever the title needs to be
changed, including when decorations need to be removed (e.g. when
move or edit mode is turned off), so that the title string doesn't
need to be saved/restored before/after move or edit.

For backward compatibility, the value of self.saved_title is
maintained, but it is not used by the new code.

The leading space in the mode and edit mode strinhs is now consistently done in the style of
' ' + _("(TV)")
rather than having the space embedded in the translatable strings
in some places and separate in others.

The order of the inheritance of the ChannelSelectionEdit mixin has
been changed so that its titleStringDecoration() overrides the same
method name in ChannelSelectionBase.