[skin.py] Fix a few hard coded skin attributes
Change the "scale", "enableWrapAround" and "noWrap" skin tag attributes so that skin provided values are no longer ignored and replaced with hardcoded default values.  Some validation is done of the attribute values.  The old values "0" and "1" are still valid.  Skinners can now also use case insensitive "Yes", "Enabled", "On", "True", "Yes" and the attribute name as alternate values to indicate that the attribute is being asserted.  Any other values will be considered that the attribute is being negated.