[ChannelSelection] allow to delete an item in move mode
This commit allows to delete channel item in move mode.
Previously it was also possible to delete an item by pressing 4.
Although when item was selected a crash was occuring.
So we are display the delete and we are making sure crash won't happen.

Note: merged from OpenPLi with conficts (that required changes) so needs testing.