[Renderer.Label] Akkow Label to be used with non-text Sources
Fixes a bug with Renderer.Lable when it tries to access source.text
whan the upstream has no text attribute.

For example, in ViX-Common (used by ViX-Night-HD and ViX-Day-HD
skins) widgetif you use this:

<widget source="VKeyIcon" conditional="VKeyIcon" render="Label" text="TEXT" position="1200,686" zPosition="1" size="35,25" transparent="1" alphatest="on">
	<convert type="ConditionalShowHide"/>
</widget>

instead of the corresponding widget with a Pixmap renderer in
FullscreenConfigPanel, the TEXT label never displays and an exception
is thrown in the skin code "Error in screen 'Setup': 'ConditionalShowHide'
object has no
 attribute 'text'".

Change Label.changed() to only try to access self.source.text if
the text attribute exists.