Multiboot - this uses code from OpenPli to improve the overall execution - and also uses it in the Emptyslot routine (#534)
* Console

* Multiboot: use tempfile to create temporary directiries

* self should not be used in a function

also, tmp_mount should be tmp.dir

* For multiboot do not rmtree to avoid 'softbrick' boxes

When by accident the partition is still mounted it could remove full
images and/or the STARTUP files that 'softbrick' the boxes. You need to
reinstall images and/or the recovery image. This avoids it in case
something odd happens. Then better get a GSOD instead of 'softbrick' a
box

* Multiboot - cleanup OpenPli merge

* Multiboot - improve EmptySlot routine

Co-authored-by: Tony Whalley <awhalley@waitrose.com>
Co-authored-by: littlesat <littlesat99@yahoo.com>
Co-authored-by: Harro Verton <wanwizard@wanwizard.eu>