[UpdateChoices] ActionMap fixup after ChoiceBox changes
Reduce the actionmap in UpdateChoices to the minimum needed by the
class, and rename it from "actions" to "menuActions" so that it
doesn't overwrite the "actions" ActionMap in the parent class.

This fixes the break of the EXIT button in this class introduced
in commit 2811605, which merged the "back" action in the ChoiceBox
"cancelaction" ActionMap into the ChoiceBox "actions" ActionMap.

Because that action wasn't replicated in the UpdateChoices "actions"
ActionMap, the overwrite of the parent's ActionMap cause the "back"
action to be lost.