[MovieSelection] Refactor MovieBrowserConfiguration to use Setup
Refactor MovieBrowserConfiguration to use Setup and setup.xml instead
of using ConfigListScreen and generating the config list in the
class.

There are two minor UI changes:

The original code used OK and MENU as alternatives for save and
cancel. Those buttons are used for other useful functions in recent
ConfigListScreens, and so they are no longer used for save/exit.

The green and red hint changes from "OK" to "Save" in line with
other Setup screens.