[Screen.py] Support display skin titles
This changes adds a "title" widget that is used buy some display skins via "parent.title".  This widget is *not* the same as the GUI "Title" as this widget only contains the current screen title and not the screen path even if the GUI title includes the screen path.

The "title" widget has the same value as the display screen "Title" widget.