[(Power)TimerEntry] Make UI closer to standard Setup screens
Remove "select" ActionMap entry so that it reverts to its standard
uses in ConfigText items.

Have keyCancel() call ConfigListScreen.closeConfigList() with an
appropriate argument so that it remains compatible with the callback
in (Power)TimerList, but still warns the user about cancelling after
changes.

[TimerEntry]

Remove keyLeft() and keyRight() methods so that they revert to their
standard uses in ConfigText items.