[ChannelSelection] Use stable enable/disable for self['dialogactions']
In ChannelSelectionEPG, self['dialogactions'] is enabled/disabled
using execBegin/End().

This is defeated by the fact that when execDialog() is called on
ChannelSelection, execBegin() is called on all ActionMaps in the
screen, including on self['dialogactions'].

This resulted in the cancel action in both self['actions'] and
self['dialogactions'] being active at the same time.

In some images, this resulted in two presses of EXIT being needed
to exit ChannelSelection.

This fix replaces calls to self['dialogactions'].execBegin/End()
with calls to self['dialogactions'].etEnabled(False/True).

Disabling an ActionMap using setEnabled() is not affected by calls
to execBegin().