[ChannelSelection,EPGSelectionBase] Disable help in popup menus
From live TV: UP, REC, HELP and EPG, REC, HELP display the main screen's help when a temporary popup menu is being shown.

In the case of ChannelSelection (UP, REC, HELP), it also locks up
the UI so that EXIT cannot be used (but using OK exits the popup,
then EXIT exits the help screen, but that also sets a timer).

In both cases, disable self["helpActions"] while the
popup is showing.