[InfoBarGenerics] Restore ffw/rew in timeshift viewing & remove unused actions
The entries for the seekFwd and seekBack actions in self["SeekActionsPTS"]
in InfoBarSeek were inadvertently removed when the recent help
improvements were made in InfoBarGenerics.

Restore the actions to the ActionMap.

Also remove the inadvertantly added actions on seekFwdManual,
seekBackManual, SeekbarFwd & SeekbarBack, which have no entries in
the ActionMap's context.