[ConfigList.py] Adjust ActionMap initialisation
This change adjusts the initialisation of the "setEnabled()" method of some of the ActionMaps depending on the status of the "fullUI" flag.  This should address any code that uses copies of the ConfigList code that is not up to the current standards in the OpenViX version of ConfigList.py.  The issue is that when the older versions of the "handleInputHelpers()" method is used the older code is not correctly managing the new ActionMaps.  This change should address that issue without negatively impacting OpenViX users.