[HelpMenuList] Sort help list flat alphabetic... The sort is by the description of the action (main description and extended description).
Sort is not simple because some items are strings, some lists and some callable objects. Callable objects are called prior to the sort so the the sort is carried out on the return value of the callable object.

If the sort contains one or more identical descriptions, e.g. "Number or SMS style data entry", these will be sub sorted based on "action", e.g. "1", "2", "3", etc in the case of numeric keys.