filepush: fix second bug in filepush patch.
When (p)poll returns "1" that can either mean data can be read or that
an exception occured on the fd. We always assumed the first, which
could lead to another dead lock:

- exception occurred on the fd (e.g. because the parent closed it
  prematurely, this seems to be a valid path in Enigma)
- read() was called and was stuck there because there was actually no
  data ready to be read.

Now the revents field is checked whether it actually contains a POLLIN
event. If not, assume an error and handle it like it was thrown by the
call to read().