encoder: stop playing service when PMT/VPID/APID are known and transcoding file.
The playing service puts the file data in the pvr device and is then
fetched from the demux device. This does not work for Broadcom
transcoding, we pump the data from the file directly to the encoder, so
the playing service, pumping data around, is not required when all
relevant data is known.