Use width and height instead of size in LoadPixmap
There are many places in enigma where the pixmap is loaded in python code instead of in the skin and the size of the widget depends on size of the loaded image, but cannot be specified in the skin.
Therefore use the width and height attributes because these dimensions can be specified in the python code and based on them you can scale the svg image.